requirements.txt
requests
supervision
pillow
transformers
accelerate
yt_dlp
moviepy==1.0.3